move module entpoints and modules out of control and into their own python files.
