+++ robot-server/robot_server/service/main.py
@@ -2,7 +2,8 @@ from opentrons import __version__
 from fastapi import FastAPI
 from starlette.responses import JSONResponse
 from starlette.requests import Request
-from .routers import health, networking, control, settings, deck_calibration
+from .routers import health, networking, control, settings, deck_calibration, \
+    modules
 from .models import V1BasicResponse
 from .exceptions import V1HandlerError
 
@@ -28,6 +29,8 @@ app.include_router(router=settings.router,
                    tags=["settings"])
 app.include_router(router=deck_calibration.router,
                    tags=["deckCalibration"])
+app.include_router(router=modules.router,
+                   tags=["modules"])
 
 
 @app.exception_handler(V1HandlerError)
